requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose\[cryptography\]
passlib\[bcrypt\]
python-multipart
bcrypt==4.0.1
